Use single curl command instead of two.
Let curl tell us if it successfully downloaded the file with the error code.  We
pass the -f flag to curl and it will always return a non-zero exit code unless
the download was successfull, we can simply check this code directly rather than
parsing the status code of the http request.

Also pass the -L flag to curl so it follows redirects properly.

We don't need to track if there was a successfull download but instead just
return from the function right away on success.  We can then assume that if the
loop completes and falls through there was no file downloaded and we can spit
out the error message and exit at that point.