Use bash regexes for line matching.
Instead of relying on the number of space-separated fields to match fields in
the sources lines and then having to further parse each field out with multiple
external commands use bash regexes to both match the line and parse the fields
in one shot without calling out to external commands.

Use an array to map the hash string length to the hash type instead of a case
statement.

Don't bother getting the hash type if we already have it from the sources line.